pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jenkins-log-sentinel"
version = "0.1.0"
description = "Periodic credential leak detector for Jenkins build logs"
requires-python = ">=3.11"
dependencies = [
    # Scanner
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "pyyaml>=6.0",
    "httpx>=0.27",          # Ollama REST client
    # Database
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "psycopg2-binary>=2.9",
    # API
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    # Notifications
    "apprise>=1.9",
    # Utilities
    "python-dateutil>=2.9",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "httpx>=0.27",          # TestClient for FastAPI
    "factory-boy>=3.3",
]

[project.scripts]
sentinel-scan = "scanner.main:cli"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
]

[tool.hatch.build.targets.wheel]
packages = ["scanner", "api", "db", "notifications"]
